Move test/util to test/test-util
This allows to test files in src/util
